+++ src/threads.js
@@ -65,7 +65,7 @@ StagePickerMorph, CustomBlockDefinition*/
 
 /*jshint esversion: 11, bitwise: false, evil: true*/
 
-modules.threads = '2022-May-25';
+modules.threads = '2022-May-27';
 
 var ThreadManager;
 var Process;
@@ -5639,6 +5639,7 @@ Process.prototype.doSetBlockAttribute = function (attribute, block, val) {
         oldSpec,
         expr,
         def,
+        inData,
         template,
         oldType,
         type;
@@ -5655,9 +5656,13 @@ Process.prototype.doSetBlockAttribute = function (attribute, block, val) {
         if (def.isGlobal) {
             return ide.sprites.asArray().concat([ide.stage]).some((any, idx) =>
                 any.usesBlockInstance(def, false, idx)
+            ) || ide.stage.allBlockInstancesInData(def).some(any =>
+                !any.isUnattached()
             );
         }
-        return rcvr.allDependentInvocationsOf(oldSpec).length > 0;
+        return rcvr.allDependentInvocationsOf(oldSpec).some(any =>
+            !any.isUnattached()
+        );
     }
 
     function remove(arr, value) {
@@ -5733,17 +5738,23 @@ Process.prototype.doSetBlockAttribute = function (attribute, block, val) {
         type = +val;
         if (type === 1 && !def.isGlobal) {
             // make global
+            inData = ide.stage.allContextsInvoking(def.blockSpec(), rcvr);
             def.isGlobal = true;
             remove(rcvr.customBlocks, def);
             ide.stage.globalBlocks.push(def);
         } else if (type === 2 && def.isGlobal) {
             // make local
+            inData = ide.stage.allContextsUsing(def);
             def.isGlobal = false;
             remove(ide.stage.globalBlocks, def);
             rcvr.customBlocks.push(def);
         } else {
             return;
         }
+        inData.forEach(context => {
+            context.expression = def.blockInstance();
+            context.changed();
+        });
         break;
     default:
         return;