+++ HISTORY.md
@@ -49,6 +49,9 @@
 * **Translation Updates:**
     * German
 
+### 2022-05-27
+* blocks, threads: update programmatic custom block scope changes in data references
+
 ### 2022-05-25
 * threads: update programmatic custom block-type changes in data references
 * byob: update manual custom block-type changes in data references